Migrates song configure dialog step to separate component
